requirements.txt
google-adk
firebase-admin
python-dotenv
google-genai